Refactoring change: move auth level
